Add Amazon EC2 Micro instance.
In order to fix JENKINS-10331 (adding micro instance type), it was
needed to replace the old Amazon AWS library (Typica) with Amazons own
SDK for Java. Typica hasn't been updated in over a year and the project
seems to have stagnated. Now that Amazon has picked up the ball we might
as well use that.