Provide backwards compatibility for EC2Slave name change.
Commit 88222567e84ae42c602d5ac8430a95957ea083e3 changed the way that
Node objects are constructed from EC2Slave objects, in order to make the
Node 'name' fields more useful. It also added a field called 'instanceId'
to EC2Slave, and stopped relying on Node.getNodeName() to recover that
information. In order to provide compatibility with configurations from
before this change was made, this patch notes whether instanceId is null,
and if it is, extracts the instance ID from the Node's name.