Ensure that template instance caps are not zero.
Until this patch it was possible to explicitly set a template's
instance cap to zero; this caused no harm as template instance caps
were not enforced until version 1.18. With that version, though,
a cap of zero will keep any instances from being launched, which is
not useful (and is somewhat difficult to debug). This patch disallows
setting template instance caps to zero, and ensures that any that are
already configured to zero are ignored.