Fix for JENKINS-15081. We store the EC2Cloud that a slave belongs to as an attribute in the EC2AbstractSlave class and have a function call to retrieve it in EC2Computer.
All references to EC2Cloud.get changed to EC2AbstractSlave.cloud or EC2Computer.getCloud().
This fix allows for us to always know the correct Cloud that an EC2AbstractSlave or EC2Computer belongs to so we can have multiple EC2Clouds configured on one Jenkins instance and have different templates for different clouds and have the instances for each spin up properly.