Windows slave support
This changeset introduces support for windows AMI to the ec2-plugin.
The windows slaves are connected to with SMB over TCP to send the initial
slave.jar and connected to from jenkins using WinRM/WinRS.

Unlile linux based AMI, the windows AMI requires

* a specific security group allowing SMB over TCP (incoming TCP port 445)
* a specific securify group alloaing WinRM (incoming TCP port 5985)
* a java virtual machine installed in the Windows %PATH%
* SMB over TCP should be enabled in the windows firewall (this can be
done with: netsh firewall set portopening tcp 445 smb enable)
* WinRM must be configured and enabled, with the following commands:
  + winrm quickconfig (answer y to both questions)
  + winrm set winrm/config/service/Auth @{Basic="true"}
  + winrm set winrm/config/service @{AllowUnencrypted="true"}
  + winrm set winrm/config/winrs @{MaxMemoryPerShellMB="1024"}
* a fixed administrator password (there's no way for the plugin
to figure out the random password of the base EC2 Windows AMI)

This changeset enhance the Slave Template configuration to specify a
Windows AMI. In which case it is necessary to enter the correct
windows administrator password of the configured remote user.

Note1: the current WinRM implementation is vaguely based on Xebialabs'
Overthere and WinRB ruby implementation.

Note2: for https support (and thus encrypted communication) more AMI setup
is necessary (see the inline plugin help for more information)

This feature is considered experimental (and even more than that for the https
transport), even though it has been successfully tested,
I'm not running this yet on production.

Signed-off-by: Brice Figureau <brice@daysofwonder.com>