Wait indefinitely when launching slave agent with a timeout of zero
Summary:
The docs claim that a timeout of zero means forever, but sometimes it
turns out to really mean zero. We correct that here.

Test Plan:
Ran the new code on our Jenkins installation, saw that we didn't get the
timeout message where Jenkins waits for only about 5 seconds before
claiming that it failed to launch the slave agent.

Reviewers: csilvers

Reviewed By: csilvers

Differential Revision: http://phabricator.khanacademy.org/D10554