JENKINS-20742: Handle correct AWS error when updating tags
As per the AWS API documentation[1] the error code
"InvalidInstanceRequestID.NotFound" does not exist. I believe this was a
copy/paste mistake when adding retry support for updating tags for
on-demand instances, as evidence by the following commits: ccf31c3[2],
9a596f0[3].

I've changed the error code to now handle the correct error code,
"InvalidInstanceID.NotFound" which is the same error code that
EC2Computer._describeInstance()[4] handles when retrying.

1. http://docs.aws.amazon.com/AWSEC2/latest/APIReference/errors-overview.html
2. https://github.com/jenkinsci/ec2-plugin/commit/ccf31c36a10e324472bf6218f9387287679359a2#diff-f2115e33148d3db7c133fe014ad9dfdd
3. https://github.com/jenkinsci/ec2-plugin/commit/9a596f0577b29a3e1835143f5d51520babdd7c1f#diff-f2115e33148d3db7c133fe014ad9dfdd
4. https://github.com/jenkinsci/ec2-plugin/blob/ec2-1.26/src/main/java/hudson/plugins/ec2/EC2Computer.java#L143