Add the ability to programatically update the AMI ID using Groovy
This is a proposed fix for JENKINS-28268

Our Use Case:

At FreshBooks, we have multiple Packer jobs that build new Jenkins build
workers as and when the underlying repos are updated. The problem with this
approach (currently) is that after this is done, someone needs to go in and
manually update the Jenkins configuration to point to the new AMI ID.

Proposed Fix:

With this approach, we could trigger a post-build Groovy script to
automatically go in and update the relevant AMI ID without us having to do this
manually.

Groovy Example:

Here is an example Groovy script that demonstrates getting and setting the AMI
ID programatically:

```
import jenkins.model.*;
import hudson.plugins.ec2.*;

Jenkins.instance.clouds.each {
  if (it.displayName == 'test-cloud') {
    it.getTemplates().each {
      println(it.displayName)
      println("Print current AMI id : " + it.getAmi())
      it.setAmi("ami-4567")
      println("Print current AMI id : " + it.getAmi())
    }
  }
}

Jenkins.instance.save()
```

Reference:

https://issues.jenkins-ci.org/browse/JENKINS-28268

[FIXED JENKINS-28268]