JENKINS-25832: Launch multiple slaves in parallel
This commit changes the behaviour of core functionality: raising new
nodes. It implements very simple algorithm:
- if NodeProvisioner requested X executors, start to raise
X/N nodes (N is number of executors on each node).
- if there is any orphant nodes (~non known by Jenkins at all), take
them in account (and start if it's neccessary)
- if there not enough orphant nodes - raise needed nodes in parallel
- give them back to NodeProvisioner in connected state
- nodes are connected by Jenkins itself: nodes are not connected
explicitly anymore (as it's needed by NodeProvisioner contract)
- EC2Cloud always returns instance in valid (RUNNING) state, it allows
to simplify logic of launcher

And reduce amount of duplicated code a little bit.