[JENKINS-53322]Cross-version dom4j compatibility tweaks.
Tweaks to make ec2 compile and pass tests on both the currently-used ancient dom4j and the latest community supported version. This makes it possible to remove our reliance of long outdated fork.

This is one piece of the components required for JENKINS-53322, but sufficient to make the ec2 plugin work on old or new.