[JENKINS-62043] Fix NPE which broke Jenkins builds
In PR #433, which was released in AWS EC2 Plugin 1.50, a bug was introduced due to improper handling of an
important transient field which is accessed very frequently by the Jenkins server due to it being referenced in
`EC2AbstractSlave.isAcceptingTasks()`. This broke any Jenkins user which had running EC2 Nodes during the
update to plugin 1.50 (see https://issues.jenkins-ci.org/browse/JENKINS-62043 for details).