[JENKINS-62043] Fix NPE which broke Jenkins builds (#452)
* [JENKINS-62043] Fix NPE which broke Jenkins builds

In PR #433, which was released in AWS EC2 Plugin 1.50, a bug was introduced due to improper handling of an
important transient field which is accessed very frequently by the Jenkins server due to it being referenced in
`EC2AbstractSlave.isAcceptingTasks()`. This broke any Jenkins user which had running EC2 Nodes during the
update to plugin 1.50 (see https://issues.jenkins-ci.org/browse/JENKINS-62043 for details).

* Excluded false-positive nullcheck Spotbugs rule

Turned off a Spotbugs rule which incorrectly understands how Java serialization works, and which thought a transient field
couldn't be null in Object.readResolve() when it actually can be.

* Removed @Nonnull annotation from transient field 

Removed a javax.annotation.Nonnull annotation from a transient field to silence a Spotbugs warning which
incorrectly flagged a field as never being null when it can in fact be null due to Java serialization semantics

* Removed unneeded NONNULL exclusion

Code in EC2AbstractSlave was updated to remove a javax.annotation.Nonnull annotation, so this Spotbugs warning should
no longer trigger and we can remove the exclusion.