Use description instead of AMI in template string
In order to make the image ID determined by properties, it can no longer
be a static string. In that case, including it in the class
representation would require that it be queried or supplied by the
caller. The AMI can be seen in the logs at other times, so just use the
description instead.