Add describable class for AWS Filter
In order to make the AWS Filter class available in configurations, add
a wrapper EC2Filter class that implements Descriptor. Ideally the values
member would be `List<String>`, but currently jenkins doesn't have a
usable way to offer those in forms. Instead the values are interpreted
as a whitespace separated list.