Raise an exception when AMI search is blank.
We use the configuration-as-code plugin; when interpolating secrets, if
you have a typo in the secret name, it inserts the empty string:

https://github.com/jenkinsci/configuration-as-code-plugin/blob/e7501727c4704aaa1e95bdc246a1454dc5b20440/plugin/src/main/java/io/jenkins/plugins/casc/SecretSourceResolver.java#L124

Unfortunately we had a typo in the name of an AMI used by a rarely-run job,
which used a slightly different AMI from all the other jobs on our server.
So, when we deployed, we missed the warning message:

    Configuration import: Found unresolved variable '/plugins/amazonEC2/special-ami'. Will default to empty string

... then a day later, the job that used this special ami ran. Because the
code warned rather than throwing an exeption, it attempted to search for an
AMI with no search parameters, then picked one of the results to launch.
As luck would have it, there was also a typo in our IAM rule that prevented
launching public AMIs; eventually, after two days of retries, the job managed
to launch one of the random AMIs: a bitcoin miner :facepalm:. The node
was shut down seconds later when jenkins failed to connect to it, and our
alarms noticed the miner, leading us to investigate where it had originated.

Hence this PR...I really don't think anyone wants to use Jenkins as a random
AMI launcher :)