JENKINS-69581: Update required java version
As of Jenkins Weekly 2.357 and LTS 2.361 the minimum required java
version is java 11.

If the command java -fullversion fails we will try to install openjdk-11.

We first try and install it using amazon-linux-extras. If this passes
then the launch continues. If it fails (for example we're launching a
linux node on a distribution other than amazon-linux we try and install
openjdk-11 using yum.

If both of these commands fail we won't launch the node.

At this point we don't explicitly check the java version, but this could
perhaps be a future addition.