Updated Spot section of README.md
Bidding was removed in 2017. Spot pricing is now set by AWS and based on long term trends in supply and demand. Its price is dynamic, however the change is much more slow and predictable. Furthermore, the termination of Spot Instances is not an issue, it's inherent in the nature of the technology and there are mechanisms to handle the terminations.