Test incremental AWS Java SDK 1.12.573-407.x
https://issues.jenkins.io/browse/JENKINS-72007 asks for support of new EC2
instance types that are available with newer versions of the AWS Java SDK.
I'm not a user of the AWS JAva SDK, but there is an incremental build
of the 1.12.573 version of the API plugin that should be usable for tests.

Testers need to install

* https://ci.jenkins.io/job/Plugins/job/aws-java-sdk-plugin/job/PR-1124/lastSuccessfulBuild/artifact/org/jenkins-ci/plugins/aws-java-sdk/aws-java-sdk-minimal/1.12.573-407.v9369a_87864ca_/aws-java-sdk-minimal-1.12.573-407.v9369a_87864ca_.hpi
* https://ci.jenkins.io/job/Plugins/job/aws-java-sdk-plugin/job/PR-1124/lastSuccessfulBuild/artifact/org/jenkins-ci/plugins/aws-java-sdk/aws-java-sdk-ec2/1.12.573-407.v9369a_87864ca_/aws-java-sdk-ec2-1.12.573-407.v9369a_87864ca_.hpi

Once those two are installed, then the incremental build of this plugin should work.