Exclude PA_PUBLIC_PRIMITIVE_ATTRIBUTE spotbugs warning
This exclusion is widely used in Jenkins plugins because public primitives
are often part of the public API of a Java object in Jenkins.

Removes the spotbugs plugin from the pom file because the parent pom
already provides it and provides additional configuration to make spotbugs
even more useful for a Jenkins plugin.

Resolves the 19 spotbugs warnings that are reported on the master branch.